[controller] re-struct core part
