+++ src/identity.rs
@@ -34,12 +34,6 @@ impl TaskId {
     pub fn get(self) -> u64 {
         self.0
     }
-
-    /// Reconstructs a `TaskId` from its raw value.
-    #[inline]
-    pub(crate) fn from_raw(raw: u64) -> Self {
-        TaskId(raw)
-    }
 }
 
 impl std::fmt::Display for TaskId {
@@ -63,12 +57,5 @@ mod tests {
     #[test]
     fn never_mints_zero() {
         assert!(TaskId::next().get() >= 1);
-    }
-
-    #[test]
-    fn from_raw_round_trips() {
-        let id = TaskId::from_raw(42);
-        assert_eq!(id.get(), 42);
-        assert_eq!(id, TaskId::from_raw(42));
     }
 }
